requirements.txt
uvicorn
fastapi
kafka-python
